hopefully fix to issue 1
added a function that checks the integrity of a pdf file. This helped me
to discover that the corruption seems to be caused by
subprocess.check_call, which seems to not wait for the completion of the
command before carrying on, as should be expected.

I replaced subprocess.check_call with subprocess.Popen + wait(), which
seems to work better. For security, I added a call to detect malformed
pdf in the case this fix doesn't prove to work 100%.