now the print generator uses the table of contents from the stub instead of using it from the python code.
This removes the need for a very dodgy hardcoded value containing the number of pages in the table of contents (typically 1 or 2). This way, this hardcoded value does no longer to be adjusted for each piece, as the table of contents is read from the stub. The code that reads the table of contents from the stub is very very messy (lots of reverse engineering to figure out how pdf stores its data), it will be cleaned very soon.